server/document-service/lambda/documentHandler/hello_world/requirements.txt
requests
fastcdc
python-fossil-delta